chore: Extract scope manager scopes
